python/requirements.txt
cloudevents==1.10.1
dapr==1.11.0
fastapi==0.104.1
grpcio==1.59.2
pydantic==2.4.2
requests==2.31.0
uvicorn==0.23.2
aiohttp==3.9.0b0
opentelemetry-sdk
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-fastapi
